Hacked a bit on the new notebook performance tests.
Tried out fancy indexing on the full set of atoms and seeing speedup
that we didn't see before. This works because the resulting AtomGroup
only stores the indexes for the rows in the master table its atoms
correspond to; it doesn't do any fancy indexing on the table until
one calls a method or property, such as names() or charges(). What's
more, for these cases it only has to slice a column of the array
to get these, returning only a copy of the underlying data and only
that data (not the whole rows), which appears to be quite fast.